refactored desktopwindowmanager
