+++ AmethystWindowsSystray/AmethystWindowsSystray.csproj
@@ -61,6 +61,10 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DesktopWindowManager.generator.cs" />
+    <Compile Include="DesktopWindowManager.draw.cs" />
+    <Compile Include="DesktopWindowManager.layouts.cs" />
+    <Compile Include="DesktopWindowManager.windows.cs" />
     <Compile Include="DesktopWindowManager.Internal\DesktopWindow.cs" />
     <Compile Include="DesktopWindowManager.Internal\Utils.cs" />
     <Compile Include="Settings.cs" />
